+++ Tests/Test-CLocalGroup.Tests.ps1
@@ -11,6 +11,7 @@ Describe 'Test-CLocalGroup' {
     BeforeEach {
         $Global:Error.Clear()
     }
+
     It 'finds local groups' {
         $groups = Get-LocalGroup
         $groups | Should -Not -BeNullOrEmpty
@@ -20,6 +21,21 @@ Describe 'Test-CLocalGroup' {
 
     It 'handles non-existent group' {
         Test-CLocalGroup -Name 'jfnrqnwuiocnja' | Should -BeFalse
+        $Global:Error | Should -BeNullOrEmpty
+    }
+
+    It 'allows wildcards' {
+        Test-CLocalGroup -Name 'Admin*' | Should -BeTrue
+        $Global:Error | Should -BeNullOrEmpty
+    }
+
+    It 'ignores wildcards' {
+        Test-CLocalGroup -LiteralName 'Admin*' | Should -BeFalse
+        $Global:Error | Should -BeNullOrEmpty
+    }
+
+    It 'finds using exact name' {
+        Test-CLocalGroup -LiteralName 'Administrators' | Should -BeTrue
         $Global:Error | Should -BeNullOrEmpty
     }